removing AssistantPreview, replacing with AssistantPreview2
